% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordler.R
\name{assess_guess}
\alias{assess_guess}
\title{Assess a guess against the target word}
\usage{
assess_guess(game)
}
\arguments{
\item{game}{Wordler game object.}
}
\value{
Character vector of guess assessment.
}
\description{
Assess a guess against the target word
}
