% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordler.R
\name{check_guess_hard_mode}
\alias{check_guess_hard_mode}
\title{Check if current guess complies with hard_mode rules}
\usage{
check_guess_hard_mode(guess, game)
}
\arguments{
\item{guess}{The guess}

\item{game}{Wordler game object.}
}
\value{
bool
}
\description{
Check if current guess complies with hard_mode rules
}
