% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordler.R
\name{count_freqs}
\alias{count_freqs}
\title{Get counts of each letter in the target}
\usage{
count_freqs(xs, target)
}
\arguments{
\item{xs, target}{we count the occurrences of each element in
\code{xs} in \code{target}}
}
\value{
Named list of elements of \code{xs} with counts.
}
\description{
Get counts of each letter in the target
}
