% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordler.R
\name{have_a_guess}
\alias{have_a_guess}
\title{Submit a guess word to the wordler game object}
\usage{
have_a_guess(x, game)
}
\arguments{
\item{x}{The guess.}

\item{game}{The wordler game object.}
}
\value{
A wordler game object.
}
\description{
Submit a guess word to the wordler game object
}
