% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordler.R
\name{is.wordler}
\alias{is.wordler}
\title{Detects wordler objects}
\usage{
is.wordler(x, ...)
}
\arguments{
\item{x}{An R object}

\item{...}{Additional arguments}
}
\value{
Returns TRUE if x is a wordler object, otherwise FALSE.
}
\description{
Detects wordler objects
}
