% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordler.R
\name{is_guess_correct}
\alias{is_guess_correct}
\title{Establish if guess is correct and set game state accordingly}
\usage{
is_guess_correct(game)
}
\arguments{
\item{game}{'wordler' game object (as generated by
\code{\link{new_wordler}}).}
}
\value{
A 'wordler' game object.
}
\description{
Compares the guess in \code{game$guess} (index from \code{game$guess_count})
with the corresponding target word in \code{game$target}. If the guess is
equal to the target, \code{game$game_won} and \code{game$game_over} are
both set to \code{TRUE}.
}
