% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{keyboards}
\alias{keyboards}
\title{Keyboard layouts for printing a wordler game at the console.}
\format{
A list of length 1.
}
\source{
\url{https://gist.github.com/cfreshman/cdcdf777450c5b5301e439061d29694c}
}
\usage{
keyboards
}
\description{
A list of keyboard layouts used to show letters known to be not in target
word, in the target word, and in the right position in the target word.
Each element must be a list having 3 items, each representing a row of a
keyboard layout.
}
\keyword{datasets}
