% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordler.R
\name{make_empty_assess_matrix}
\alias{make_empty_assess_matrix}
\title{Makes an empty matrix to hold assessment of guesses}
\usage{
make_empty_assess_matrix()
}
\value{
Empty guess assessment matrix.
}
\description{
Makes an empty matrix to hold assessment of guesses
}
