% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordler.R
\name{make_empty_guess_matrix}
\alias{make_empty_guess_matrix}
\title{Makes an empty matrix to hold guesses}
\usage{
make_empty_guess_matrix()
}
\value{
Empty guess matrix.
}
\description{
Makes an empty matrix to hold guesses
}
