% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordler.R
\name{make_new_game}
\alias{make_new_game}
\title{Make a new blank wordle game}
\usage{
make_new_game(target_words = NULL)
}
\arguments{
\item{target_words}{A character vector of potential target words for the
game. A word will be randomly selected from this vector as the target word
to be guessed. Defaults to words used by the WORDLE game online
(?wordler::wordle_answers) if not provided.}
}
\value{
A list representing the Wordle game.
}
\description{
Make a new blank wordle game
}
