% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordler.R
\name{new_wordler}
\alias{new_wordler}
\title{Constructs a new object of class "wordler"}
\usage{
new_wordler(
  target = sample(wordler::wordle_answers, 1),
  game_over = FALSE,
  game_won = FALSE,
  guess_count = 0,
  guess = lapply(1:6, function(x) unlist(strsplit("_____", ""))),
  assess = lapply(1:6, function(x) rep("not_in_word", 5)),
  keyboard = wordler::keyboards$qwerty,
  letters_known_not_in_word = character(0),
  letters_known_in_word = character(0),
  letters_known_in_position = character(0)
)
}
\arguments{
\item{target}{The target word for the game. Defaults to a random selection
from words used by the WORDLE game online (?wordler::wordle_answers) if not
provided.}

\item{game_over}{A logical indicating if the game is over. Defaults to FALSE.}

\item{game_won}{A logical indicating if the game has been won. In other
words, has the target word been correctly guessed.}

\item{guess_count}{An integer representing the number of guesses made so
far. Defaults to 0.}

\item{guess}{A list (of length 6) of character vectors (each of length 5)
representing the guesses of the target word. Each element of the list
represents one of six guesses allowed. Each guess defaults to
\code{c("_", "_", "_", "_", "_")} to represent a guess not yet made.}

\item{assess}{A list (of length 6) of character vectors (each of length 5)
representing an assessment of each letter in each guess.}

\item{keyboard}{A list (of length 3) of character vectors each representing
a row of a keyboard layout used to visualise the game by \code{print()}.
Defaults to QWERTY layout.}

\item{letters_known_not_in_word}{A character vector of letters known not to
be in the target word.}

\item{letters_known_in_word}{A character vector of letters know to be in the
target word.}

\item{letters_known_in_position}{A character vector of letters known to be
in the correct position in the target word.}
}
\value{
An object of class "wordler".
}
\description{
Returns a "wordler" object which holds the state of a wordle game as guesses
are made. The returned object will have a target word which is selected from
the default list unless provided in the \code{target_words} argument.
}
\details{
The wordler object is a list which has the following elements:

\itemize{
  \item \code{game_over} - A logical indicating if the game is over. Set to
  \code{TRUE} if either the word is correctly guessed, or all guesses are
  used.
  \item \code{game_won}
  \item \code{guess_count}
  \item \code{target}
  \item \code{guess}
  \item \code{assess}
  \item \code{keyboard}
}
}
