% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordler.R
\name{print.wordler}
\alias{print.wordler}
\title{Prints a wordler game to the console.}
\usage{
\method{print}{wordler}(x, ...)
}
\arguments{
\item{x}{'wordler' game object (as generated by
\code{\link{new_wordler}}).}

\item{...}{additional arguments}
}
\value{
No return value.
}
\description{
Prints a wordler game to the console.
}
