% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordler.R
\name{print_guesses}
\alias{print_guesses}
\title{Print current guesses to console}
\usage{
print_guesses(guess_matrix, assess_matrix)
}
\arguments{
\item{guess_matrix}{Matrix of guesses.}

\item{assess_matrix}{Matrix of assessment of the guesses.}
}
\description{
Print current guesses to console
}
