% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordler.R
\name{print_instructions}
\alias{print_instructions}
\title{Prints instructions to play a wordler game in the console}
\usage{
print_instructions()
}
\value{
No return value.
}
\description{
Prints instructions to play a wordler game in the console
}
