% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ubuntu_dict}
\alias{ubuntu_dict}
\title{All five-letter words from the Ubuntu dictionary.}
\format{
A character vector of length 2488:
}
\source{
\url{https://ubuntu.com/}
}
\usage{
ubuntu_dict
}
\description{
A dataset containing all five-letter words from Ubuntu dictionary
`/usr/share/dict/words`.
}
\keyword{datasets}
