% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordler.R
\name{update_letters_known_in_position}
\alias{update_letters_known_in_position}
\title{Establish which letters are known to be in the correct position in the target
word}
\usage{
update_letters_known_in_position(game)
}
\arguments{
\item{game}{A wordler game object.}
}
\value{
A wordler game object.
}
\description{
Establish which letters are known to be in the correct position in the target
word
}
