% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordler.R
\name{update_letters_known_not_in_word}
\alias{update_letters_known_not_in_word}
\title{Establish which letters are known to _not_ be in the target word}
\usage{
update_letters_known_not_in_word(game)
}
\arguments{
\item{game}{A wordler game object.}
}
\value{
A wordler game object.
}
\description{
Establish which letters are known to _not_ be in the target word
}
