% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wordle_answers}
\alias{wordle_answers}
\title{All words used as potential answers by the original WORDLE game.}
\format{
A character vector of length 2315.
}
\source{
\url{https://gist.github.com/cfreshman/a03ef2cba789d8cf00c08f767e0fad7b/}
}
\usage{
wordle_answers
}
\description{
A dataset containing all words which can be used as answers to the original
WORDLE game.
}
\keyword{datasets}
