% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordler.R
\name{assess_guess}
\alias{assess_guess}
\title{Assess a guess against the target word}
\usage{
assess_guess(guess, word)
}
\arguments{
\item{guess}{The guess. Character vector of length 1.}

\item{word}{The target word. Single row character matrix.}
}
\value{
Character vector of guess assessment.
}
\description{
Assess a guess against the target word
}
