% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordler.R
\name{have_a_guess}
\alias{have_a_guess}
\title{Submit a guess word to a wordler game object}
\usage{
have_a_guess(x, game, allowed_words = NULL)
}
\arguments{
\item{x}{the guess.}

\item{game}{'wordler' game object (as generated by
\code{\link{new_wordler}}).}

\item{allowed_words}{a character vector of valid words for the guess. x
must be in this vector to be allowed. Defaults to words used by the WORDLE
game online (?wordler::wordle_allowed) if not provided.}
}
\value{
A 'wordler' game object.
}
\description{
If \code{x} is a valid guess, it is added to \code{game$guess} and assessed
against the target word. Increments game$guess_count if a valid guess is made.
}
