% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordler.R
\name{is.wordler}
\alias{is.wordler}
\title{Detects wordler objects}
\usage{
is.wordler(x, ...)
}
\arguments{
\item{x}{an R object}

\item{...}{additional arguments}
}
\value{
Returns \code{TRUE} if x is a 'wordler' object, otherwise
\code{FALSE}.
}
\description{
Detects wordler objects
}
