% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordler.R
\name{is_guess_correct}
\alias{is_guess_correct}
\title{Establish if guess is correct and set game state accordingly}
\usage{
is_guess_correct(game)
}
\arguments{
\item{game}{A wordler game object.}
}
\value{
A wordler game object.
}
\description{
Establish if guess is correct and set game state accordingly
}
