% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordler.R
\name{make_new_game}
\alias{make_new_game}
\title{Make a new blank wordle game}
\usage{
make_new_game(hard_mode = FALSE)
}
\arguments{
\item{hard_mode}{Flag if game is in hard mode}
}
\value{
A list representing the Wordle game.
}
\description{
Make a new blank wordle game
}
