% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordler.R
\name{play_wordler}
\alias{play_wordler}
\title{Play a game of Wordle in the R console}
\usage{
play_wordler(hard_mode = FALSE)
}
\arguments{
\item{hard_mode}{Flag if hard mode should be used. In hard mode each letter
in a correct position must be in the same position in future guesses, and
any letters in the word must be used in future guesses.}
}
\description{
Play a game of Wordle in the R console
}
