% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordler.R
\name{print.wordler}
\alias{print.wordler}
\title{Prints a wordler game to the console.}
\usage{
\method{print}{wordler}(x)
}
\arguments{
\item{x}{A wordler game object.}
}
\value{
No return value.
}
\description{
Prints a wordler game to the console.
}
