% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qdap_dict}
\alias{qdap_dict}
\title{All five-letter words from the Nettalk Corpus Syllable Data Set.}
\format{
A character vector of length 2488:
}
\source{
\url{https://CRAN.R-project.org/package=qdapDictionaries/}
}
\usage{
qdap_dict
}
\description{
A dataset containing all five-letter words from the Nettalk Corpus Syllable
Data Set as returned by qdapDictionaries::dictionaries().
}
\keyword{datasets}
