% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordler.R
\name{update_letters_known_in_position}
\alias{update_letters_known_in_position}
\title{Establish which letters are known to be in the correct position in the target
word}
\usage{
update_letters_known_in_position(game)
}
\arguments{
\item{game}{'wordler' game object (as generated by
\code{\link{new_wordler}}).}
}
\value{
A 'wordler' game object.
}
\description{
For all items in \code{game$guess}, establishes the letters which are now
known to be in the correct position in the target word. These are present as
a character vector in \code{game$letters_known_in_position} in the returned
object.
}
